core: eliminate all 51 -Wconversion warnings + add CI regression gate (task #3)
Make every libbcb core source -Wconversion clean by adding explicit casts at the
implicit integer-conversion sites. Each is a value that is provably in range
(context lengths 0..24, byte values, hash indices masked to table size, array
indices >= 0), so the casts document the existing conversion rather than change
it. Classification: all 51 were safe conversions (no real overflow/truncation
bug), made explicit:

  bt_model.c (10), btv3.c (22), bcb_prior.c (16), ce_compress.c (2), bcb_api.c (1)
    - ulong hash-mask results -> unsigned int index (value < table size)
    - int context-length -> size_t in memcpy/memcmp, -> unsigned char ctx_len
    - size_t pool index -> int slot value; int loop var -> size_t corpus index
    - int bit -> uint32_t range-coder code; CRC byte extraction -> uint8_t
    - uint64_t LUT term: unify operand ranks to avoid spurious sign-conversion

Hot path (log-domain integer math, MCU) stays integer/libm-free; no float
introduced. Verified BIT-IDENTICAL: a prior built + a message compressed before
and after the change are byte-for-byte equal (sha256 match); make test / api-test
/ threads-test / prior-equiv all pass lossless.

Regression gate: new 'make wconv' target compiles each core file with
-Wconversion -Werror; wired into .github/workflows/msgbench.yml so a reintroduced
implicit conversion fails CI.

+++ src/v0_baseline/ce_compress.c
@@ -129,7 +129,7 @@ CecDecoder *cec_dec_new(const uint8_t *data, size_t len, const CecBT *bt) {
     d->low = 0; d->high = 0xFFFFFFFFu;
     d->bt = bt;
     for (int i=0; i<32; i++)
-        d->code = (d->code << 1) | cec_read_bit_(d);
+        d->code = (d->code << 1) | (uint32_t)cec_read_bit_(d);
     return d;
 }
 
@@ -149,7 +149,7 @@ static void cec_renorm_dec_(CecDecoder *d) {
         } else break;
         d->low <<= 1;
         d->high = (d->high << 1) | 1;
-        d->code = (d->code << 1) | cec_read_bit_(d);
+        d->code = (d->code << 1) | (uint32_t)cec_read_bit_(d);
     }
 }
